requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.25.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
joblib>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0
scipy>=1.11.0
xgboost>=2.0.0
